pyproject.toml
[tool.poetry]
name = "gemini-workshop"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
google-api-python-client = "^2.133.0"
pillow = "^10.3.0"
opencv-python = "^4.10.0.82"
matplotlib = "^3.9.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"
rich = "^13.7.1"
jupyter = "^1.0.0"
google-generativeai = "<0.6.0"
alive-progress = "^3.1.5"
diskcache = "^5.6.3"
validators = "^0.28.3"
langchain = "^0.2.3"
chromadb = "^0.5.0"
langchain-google-genai = "^1.0.6"
faiss-cpu = "^1.8.0"
python-docx = "^1.1.2"
langchain-community = "^0.2.4"
unstructured = "^0.14.5"
langchain-google-vertexai = "^1.0.5"
pandas = "^2.2.2"
scikit-learn = "^1.5.0"
umap-learn = "^0.5.6"
scipy = "^1.13.1"
tensorflow-metal = "^1.1.0"
notebook = "^7.2.1"
librosa = "^0.10.2.post1"
dlib = "^19.24.4"
imutils = "^0.5.4"
numpy = "^1.26.4"
gradio = "^4.36.1"
pydantic = "^2.7.3"
tensorflow = "^2.16.1"
google-cloud-aiplatform = {extras = ["reasoningengine"], version = "^1.51.1"}
sentence-transformers = "^3.0.1"
seaborn = "^0.13.2"
weaviate-client = "^4.6.4"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
